Deduplicate shared web helpers
