--- web/apps/albums/src/public-album/media/processing/ffmpeg/constants.ts
@@ -1,3 +1,0 @@
-export const ffmpegPathPlaceholder = "FFMPEG";
-export const inputPathPlaceholder = "INPUT";
-export const outputPathPlaceholder = "OUTPUT";
