+++ web/apps/albums/src/public-album/upload/pipeline/thumbnail.ts
@@ -1,5 +1,5 @@
-import * as ffmpeg from "@/public-album/media/processing/ffmpeg";
 import log from "ente-base/log";
+import * as ffmpeg from "ente-gallery/services/ffmpeg";
 import { FileType, type FileTypeInfo } from "ente-media/file-type";
 import { isHEICExtension } from "ente-media/formats";
 import { heicToJPEG } from "ente-media/heic-convert";
