+++ web/apps/embed/src/pages/_app.tsx
@@ -2,6 +2,7 @@ import "@fontsource-variable/inter";
 import { CssBaseline } from "@mui/material";
 import { ThemeProvider } from "@mui/material/styles";
 import { CustomHead } from "ente-base/components/Head";
+import { ThemedLoadingBar } from "ente-base/components/ThemedLoadingBar";
 import {
     useIsRouteChangeInProgress,
     useSetupI18n,
@@ -11,7 +12,6 @@ import { photosTheme } from "ente-base/components/utils/theme";
 import { useLoadingBar } from "ente-base/components/utils/use-loading-bar";
 import { BaseContext, deriveBaseContext } from "ente-base/context";
 import { logStartupBanner } from "ente-base/log-web";
-import { ThemedLoadingBar } from "ente-new/photos/components/ThemedLoadingBar";
 import { PhotosAppContext } from "ente-new/photos/types/context";
 import type { AppProps } from "next/app";
 import "photoswipe/dist/photoswipe.css";
