+++ web/packages/base/package.json
@@ -21,6 +21,7 @@
         "react": "19.2.7",
         "react-dom": "19.2.7",
         "react-i18next": "16.6.6",
+        "react-top-loading-bar": "3.0.2",
         "zod": "4.4.3"
     },
     "devDependencies": {
