+++ web/packages/gallery/utils/quick-link.ts
